Intent: Consolidate node caching under tnfr.cache for consistent coherence tracking Operators involved: [Coherence|Resonance|Self-organization] Affected invariants: [#1, #5, #7] Key changes: - Moved NodeCache and node mapping helpers into tnfr.cache and reused checksum primitives - Updated imports and helper exports with lazy glyph history proxies to avoid circular loading - Deleted tnfr.helpers.node_cache and ensured node_set_checksum honours non-storing calls Expected risks/dissonances: Potential unnoticed import regressions; mitigated via targeted pytest suite Metrics: No change in C(t), Si, νf; structural caching validated by tests Equivalence map: tnfr.helpers.node_cache.* -> tnfr.cache.*
